backend/requirements.txt
fastapi==0.100.0
uvicorn==0.23.0
python-telegram-bot==20.4
tonsdk==1.0.15
sqlalchemy==2.0.20
alembic==1.11.3
psycopg2==2.9.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
websockets==11.0.3
python-dotenv==1.0.0
